# packages/contracts/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "Reference Video Studio API", "version": "v1" },
  "components": {
    "schemas": {
...
      "SafeErrorEnvelope": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": { "type": "string" },
              "message": { "type": "string" },
              "causeCategory": { "type": "string" },
              "remediation": { "type": "string" },
              "docsUrl": { "type": "string" },
              "correlationId": { "type": "string" },
              "details": { "type": "array", "items": { "type": "object" } },
              "safePredecessor": {
                "type": "object",
                "properties": {
                  "sceneVersion": { "type": "integer", "minimum": 0 },
                  "sceneDigest": { "type": "string" },
                  "artifactId": { "type": "string" }
                },
                "additionalProperties": false
              }
            },
            "required": [
              "code",
              "message",
              "causeCategory",
              "remediation",
              "docsUrl",
              "correlationId",
              "details"
            ],
            "additionalProperties": false
          }
        },
        "required": ["error"],
        "additionalProperties": false
      },
...
      "MotionSceneRollbackV1": {
        "type": "object",
        "properties": {
          "schema": { "type": "string", "const": "motion-scene-rollback-v1" },
          "version": { "type": "integer", "minimum": 1 }
        },
        "required": ["schema", "version"],
        "additionalProperties": false
      },
      "MotionSceneRenderV1": {
        "type": "object",
        "properties": {
          "schema": { "type": "string", "const": "motion-scene-render-v1" },
          "backend": { "type": "string", "enum": ["native", "adobe"] },
          "deviceId": { "type": "string" },
          "projectId": { "type": "string" }
        },
        "required": ["schema"],
        "additionalProperties": false
      },
      "MotionSceneSnapshotV1": {
        "type": "object",
        "properties": {
          "schema": { "type": "string", "const": "motion-scene-snapshot-v1" },
          "version": { "type": "integer", "minimum": 1 },
          "sceneEtag": { "type": "string" },
          "sceneDigest": { "type": "string", "pattern": "^[a-f0-9]{64}$" },
          "scene": { "type": "object" },
          "history": { "type": "array", "items": { "type": "object" } },
          "backendCapability": { "type": "object" },
          "verification": { "type": ["object", "null"] },
          "planDigest": {
            "type": ["string", "null"],
            "pattern": "^[a-f0-9]{64}$"
          },
          "predecessorVersion": { "type": ["integer", "null"], "minimum": 1 },
          "artifactDigest": {
            "type": ["string", "null"],
            "pattern": "^[a-f0-9]{64}$"
          },
          "predicateIds": { "type": "array", "items": { "type": "string" } },
          "knowledgeCardIds": {
            "type": "array",
            "items": { "type": "string" },
            "maxItems": 15
          },
          "knowledgeCards": {
            "type": "array",
            "maxItems": 15,
            "items": {
              "type": "object",
              "properties": {
                "id": { "type": "string" },
                "domain": { "type": "string" },
                "titleEn": { "type": "string" },
                "titleKo": { "type": "string" }
              },
              "required": ["id", "domain", "titleEn", "titleKo"],
              "additionalProperties": false
            }
          },
          "adobeDevices": {
            "type": "array",
            "maxItems": 32,
            "items": {
              "type": "object",
              "properties": {
                "id": { "type": "string" },
                "label": { "type": "string" }
              },
              "required": ["id", "label"],
              "additionalProperties": false
            }
          },
          "adobeProjects": {
            "type": "array",
            "maxItems": 32,
            "items": {
              "type": "object",
              "properties": {
                "id": { "type": "string" },
                "label": { "type": "string" }
              },
              "required": ["id", "label"],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "schema",
          "version",
          "sceneEtag",
          "sceneDigest",
          "scene",
          "history",
          "backendCapability",
          "verification",
          "planDigest",
          "predecessorVersion",
          "artifactDigest",
          "predicateIds",
          "knowledgeCardIds",
          "knowledgeCards",
          "adobeDevices",
          "adobeProjects"
        ],
        "additionalProperties": false
      },
...
    },
    "securitySchemes": { "bearerAuth": { "type": "http", "scheme": "bearer" } }
  },
  "paths": {
...
    "/v1/jobs/{id}/motion-scene/rollback": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": { "type": "string" }
        }
      ],
      "post": {
        "operationId": "rollbackMotionScene",
        "parameters": [
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MotionSceneRollbackV1" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "New version restored from history",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MotionSceneSnapshotV1"
                }
              }
            }
          },
          "400": {
            "description": "Request rejected",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SafeErrorEnvelope" }
              }
            }
          },
          "404": {
            "description": "Request rejected",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SafeErrorEnvelope" }
              }
            }
          },
          "409": {
            "description": "Request rejected",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SafeErrorEnvelope" }
              }
            }
          },
          "422": {
            "description": "Request rejected",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SafeErrorEnvelope" }
              }
            }
          },
          "428": {
            "description": "Request rejected",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SafeErrorEnvelope" }
              }
            }
          }
        }
      }
    },
    "/v1/jobs/{id}/motion-scene/render": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": { "type": "string" }
        }
      ],
      "post": {
        "operationId": "renderMotionScene",
        "parameters": [
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MotionSceneRenderV1" }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Queued",
            "content": {
              "application/json": { "schema": { "type": "object" } }
            }
          },
          "400": {
            "description": "Request rejected",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SafeErrorEnvelope" }
              }
            }
          },
          "404": {
            "description": "Request rejected",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SafeErrorEnvelope" }
              }
            }
          },
          "409": {
            "description": "Request rejected",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SafeErrorEnvelope" }
              }
            }
          },
          "422": {
            "description": "Request rejected",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SafeErrorEnvelope" }
              }
            }
          },
          "428": {
            "description": "Request rejected",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SafeErrorEnvelope" }
              }
            }
          }
        }
      }
    },
...
  }
}